Use podman cli way of pod start (#167)
- With SDK way of starting pod as part of `ai-services application
start` flow, we are seeing both vLLM and reranker containers getting
crash with error `The Linux VFIO kernel module could not create the TCE
window for the sPAPR IOMMU. This is an unexpected condition`. However,
when switch to podman cli way of starting pod, it seemless works, both
vllm and reranker containers successfully booted.
We are sticking to this cli way of starting pod until the root cause of
why SDK way of starting pod resulting in crash is identified.

Signed-off-by: Manjunath-A-C <manjunath.a.c@ibm.com>
Co-authored-by: Dharaneeshwaran Ravichandran <17947543+dharaneeshvrd@users.noreply.github.com>